# Agent-Study/src/main/resources/swagger/openAI-txt.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Agent Study API",
    "description": "OpenAPI documentation powered by springdoc-openapi",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:6363",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/mcp/update-server": {
      "put": {
        "tags": [
          "MCP数据库管理"
        ],
        "summary": "更新MCP服务器",
        "description": "更新已存在的MCP服务器配置",
        "operationId": "updateMCPServer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBean"
                }
              }
            }
          }
        }
      }
    },
...
    "/mcp/add-server": {
      "post": {
        "tags": [
          "MCP数据库管理"
        ],
        "summary": "添加MCP服务器",
        "description": "添加新的MCP服务器配置",
        "operationId": "addMCPServer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBean"
                }
              }
            }
          }
        }
      }
    },
...
    "/mcp/server/{id}/toggle": {
      "patch": {
        "tags": [
          "MCP数据库管理"
        ],
        "summary": "启用/禁用MCP服务器",
        "description": "切换MCP服务器的启用状态",
        "operationId": "toggleMCPServer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBean"
                }
              }
            }
          }
        }
      }
    },
...
    "/sse/status": {
      "get": {
        "tags": [
          "SSE连接管理"
        ],
        "summary": "检查连接状态",
        "description": "查询指定用户的SSE连接状态",
        "operationId": "checkStatus",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/sse/connect": {
      "get": {
        "tags": [
          "SSE连接管理"
        ],
        "summary": "建立SSE连接",
        "description": "客户端通过此接口建立SSE长连接，用于接收实时消息",
        "operationId": "connect",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SseEmitter"
                }
              }
            }
          }
        }
      }
    },
...
    "/mcp/server/{id}": {
      "get": {
        "tags": [
          "MCP数据库管理"
        ],
        "summary": "获取MCP服务器详情",
        "description": "根据ID获取指定MCP服务器的详细信息",
        "operationId": "getMCPServerById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBean"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "MCP数据库管理"
        ],
        "summary": "删除MCP服务器",
        "description": "根据ID删除指定的MCP服务器",
        "operationId": "deleteMCPServer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBean"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/server/{id}/test": {
      "get": {
        "tags": [
          "MCP数据库管理"
        ],
        "summary": "测试MCP服务器连接",
        "description": "测试指定MCP服务器的连接状态",
        "operationId": "testConnection_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBean"
                }
              }
            }
          }
        }
      }
    },
...
    "/sse/disconnect": {
      "delete": {
        "tags": [
          "SSE连接管理"
        ],
        "summary": "断开SSE连接",
        "description": "主动断开指定用户的SSE连接",
        "operationId": "disconnect",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "McpServerDto": {
        "required": [
          "serverName",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "服务器ID（更新时必填）",
            "format": "int32"
          },
          "serverName": {
            "type": "string",
            "description": "MCP服务器名称",
            "example": "高德地图MCP服务"
          },
          "serverUrl": {
            "type": "string",
            "description": "MCP服务器URL",
            "example": "https://mcp.amap.com"
          },
          "description": {
            "type": "string",
            "description": "服务器描述",
            "example": "提供地图相关功能的MCP服务"
          },
          "endpoint": {
            "type": "string",
            "description": "SSE/HTTP端点路径",
            "example": "/sse?key=YOUR_API_KEY"
          },
          "type": {
            "type": "string",
            "description": "服务器类型：STDIO, HTTP, SSE",
            "example": "SSE",
            "enum": [
              "STDIO",
              "HTTP",
              "SSE"
            ]
          },
          "jsonContent": {
            "type": "string",
            "description": "STDIO类型的JSON配置内容",
            "example": "{\"command\":\"npx.cmd\",\"args\":[\"chrome-devtools-mcp@latest\"]}"
          },
          "enabled": {
            "type": "boolean",
            "description": "是否启用",
            "example": true
          }
        },
        "description": "MCP服务器请求参数"
      },
      "ResultBean": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "description": "返回信息"
          },
          "data": {
            "type": "object",
            "description": "返回数据"
          }
        }
      },
...
      "SseEmitter": {
        "type": "object",
        "properties": {
          "timeout": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    }
  }
}